@@ -4,11 +4,11 @@ import android.graphics.drawable.ColorDrawable;
 import android.graphics.drawable.Drawable;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.v7.widget.RecyclerView;
 import android.util.Log;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+import androidx.recyclerview.widget.RecyclerView;
 import lombok.Setter;
 import org.rares.miner49er.R;
 import org.rares.miner49er.layoutmanager.postprocessing.ResizePostProcessor;
@@ -18,9 +18,9 @@ import org.rares.ratv.rotationaware.RotationAwareTextView;
 import java.util.ArrayList;
 import java.util.List;
 
-import static android.support.v7.widget.RecyclerView.SCROLL_STATE_DRAGGING;
-import static android.support.v7.widget.RecyclerView.SCROLL_STATE_IDLE;
-import static android.support.v7.widget.RecyclerView.SCROLL_STATE_SETTLING;
+import static androidx.recyclerview.widget.RecyclerView.SCROLL_STATE_DRAGGING;
+import static androidx.recyclerview.widget.RecyclerView.SCROLL_STATE_IDLE;
+import static androidx.recyclerview.widget.RecyclerView.SCROLL_STATE_SETTLING;
 
 /**
  * @author rares
@@ -478,7 +478,7 @@ public class StickyLinearLayoutManager
             }
 
             int to = newItemPosition == BOTTOM || newItemPosition == NONE ?
-                    lastVisiblePosition + (decoratedChildHeight * extraChildren) : getPosition(item);   // FIXME: 8/28/18 <<<<
+                    lastVisiblePosition + (decoratedChildHeight * extraChildren) : getPosition(item);   // FIXME: 8/28/18 <<<< [possible npe on get position]
 
             if (DEBUG && METHOD_DEBUG) {
                 Log.w(TAG, "drawChildren: lastVisiblePosition: " + lastVisiblePosition);
